improve how we duplicate forces
