+++ packages/custom-types/force.d.ts
@@ -3,6 +3,9 @@ import ForceData from './force-data'
 export default interface Force {
   name: string,
   forces: Array<ForceData>,
+  // TODO: we need to constrain what we store below.
+  // I _think_ we're storing a whole force. I'm sure it should be
+  // a force-id
   selectedForce: string,
   complete: boolean,
   dirty: boolean
