+++ packages/helpers/src/index.ts
@@ -12,6 +12,7 @@ export { default as convertMilliseconds } from './convert-milliseconds'
 export { default as deepCompare } from './deep-compare'
 export { default as deletePlatform } from './delete-platform'
 export { default as deepCopy } from './deep-copy'
+export { default as duplicateThisForce } from './duplicate-force'
 export { default as findAsset } from './find-asset'
 export { default as configDateTimeLocal } from './config-date-time-local'
 export { default as findPerceivedAsTypes } from './find-perceived-as-types'
